requirements.txt
streamlit==1.28.2
numpy==1.24.3
keras==2.14.0
pillow==9.5.0
joblib==1.2.0